Extract DescriptorRepository for readability
The Linker contains a big chunk of code which only responsibility is to
retrieve a Descriptor for a given FQSEN. This is actually a secondary
purpose and to make the linker more readable and testable, I extracted
this.

I am not convinced this is the correct place for the Repository, I will
check that in a followup commit, but for now it is separate and better
readable.

Another fix I need to do is to inspect and upgrade the tests for both
classes; and with this out of the way I can make the Linker even more
readable
