+++ src/phpDocumentor/Compiler/Linker/Linker.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Compiler\Linker;
 
+use ArrayAccess;
 use phpDocumentor\Compiler\CompilerPassInterface;
 use phpDocumentor\Descriptor\ClassDescriptor;
 use phpDocumentor\Descriptor\DescriptorAbstract;
@@ -29,10 +30,6 @@ use function is_array;
 use function is_object;
 use function is_string;
 use function spl_object_hash;
-use function str_replace;
-use function strlen;
-use function strpos;
-use function substr;
 use function ucfirst;
 
 /**
@@ -55,16 +52,14 @@ class Linker implements CompilerPassInterface
 {
     public const COMPILER_PRIORITY = 10000;
 
-    public const CONTEXT_MARKER = '@context';
-
-    /** @var DescriptorAbstract[] */
-    private $elementList = [];
-
     /** @var string[][] */
     private $substitutions = [];
 
     /** @var string[] Prevent cycles by tracking which objects have been analyzed */
     private $processedObjects = [];
+
+    /** @var DescriptorRepository */
+    private $descriptorRepository;
 
     public function getDescription() : string
     {
@@ -76,14 +71,15 @@ class Linker implements CompilerPassInterface
      *
      * @param string[][] $substitutions
      */
-    public function __construct(array $substitutions)
+    public function __construct(array $substitutions, DescriptorRepository $descriptorRepository)
     {
         $this->substitutions = $substitutions;
+        $this->descriptorRepository = $descriptorRepository;
     }
 
     public function execute(ProjectDescriptor $project) : void
     {
-        $this->setObjectAliasesList($project->getIndexes()->elements->getAll());
+        $this->descriptorRepository->setObjectAliasesList($project->getIndexes()->elements->getAll());
         $this->substitute($project);
     }
 
@@ -95,16 +91,6 @@ class Linker implements CompilerPassInterface
     public function getSubstitutions() : array
     {
         return $this->substitutions;
-    }
-
-    /**
-     * Sets the list of object aliases to resolve the FQSENs with.
-     *
-     * @param DescriptorAbstract[] $elementList
-     */
-    public function setObjectAliasesList(array $elementList) : void
-    {
-        $this->elementList = $elementList;
     }
 
     /**
@@ -138,14 +124,16 @@ class Linker implements CompilerPassInterface
     public function substitute($item, $container = null)
     {
         if ($item instanceof Fqsen) {
-            return $this->findAlias((string) $item, $container);
+            return $this->descriptorRepository->findAlias((string) $item, $container);
         }
 
         if (is_string($item)) {
-            return $this->findAlias($item, $container);
+            return $this->descriptorRepository->findAlias($item, $container);
         }
 
-        if (is_array($item) || ($item instanceof Traversable && !$item instanceof ProjectInterface)) {
+        if (is_array($item)
+            || ($item instanceof Traversable && $item instanceof ArrayAccess && !$item instanceof ProjectInterface)
+        ) {
             $isModified = false;
             foreach ($item as $key => $element) {
                 $isModified = true;
@@ -198,67 +186,11 @@ class Linker implements CompilerPassInterface
     }
 
     /**
-     * Attempts to find a Descriptor object alias with the FQSEN of the element it represents.
-     *
-     * This method will try to fetch an element after normalizing the provided FQSEN. The FQSEN may contain references
-     * (bindings) that can only be resolved during linking (such as `self`) or it may contain a context marker
-     * {@see CONTEXT_MARKER}.
-     *
-     * If there is a context marker then this method will see if a child of the given container exists that matches the
-     * element following the marker. If such a child does not exist in the current container then the namespace is
-     * queried if a child exists there that matches.
-     *
-     * For example:
-     *
-     *     Given the Fqsen `@context::myFunction()` and the lastContainer `\My\Class` will this method first check
-     *     to see if `\My\Class::myFunction()` exists; if it doesn't it will then check if `\My\myFunction()` exists.
-     *
-     * If neither element exists then this method assumes it is an undocumented class/trait/interface and change the
-     * given FQSEN by returning the namespaced element name (thus in the example above that would be
-     * `\My\myFunction()`). The calling method {@see substitute()} will then replace the value of the field containing
-     * the context marker with this normalized string.
-     *
-     * @return DescriptorAbstract|string|null
-     */
-    public function findAlias(string $fqsen, ?DescriptorAbstract $container = null)
-    {
-        $fqsen = $this->replacePseudoTypes($fqsen, $container);
-
-        if ($this->isContextMarkerInFqsen($fqsen) && $container instanceof DescriptorAbstract) {
-            // first exchange `@context::element` for `\My\Class::element` and if it exists, return that
-            $classMember = $this->fetchElementByFqsen($this->getTypeWithClassAsContext($fqsen, $container));
-            if ($classMember) {
-                return $classMember;
-            }
-
-            // otherwise exchange `@context::element` for `\My\element` and if it exists, return that
-            $namespaceContext = $this->getTypeWithNamespaceAsContext($fqsen, $container);
-            $namespaceMember = $this->fetchElementByFqsen($namespaceContext);
-            if ($namespaceMember) {
-                return $namespaceMember;
-            }
-
-            // otherwise check if the element exists in the global namespace and if it exists, return that
-            $globalNamespaceContext = $this->getTypeWithGlobalNamespaceAsContext($fqsen);
-            $globalNamespaceMember = $this->fetchElementByFqsen($globalNamespaceContext);
-            if ($globalNamespaceMember) {
-                return $globalNamespaceMember;
-            }
-
-            // Otherwise we assume it is an undocumented class/interface/trait and return `\My\element` so
-            // that the name containing the marker may be replaced by the class reference as string
-            return $namespaceContext;
-        }
-
-        return $this->fetchElementByFqsen($fqsen);
-    }
-
-    /**
      * Returns the value of a field in the given object.
      *
      * @return string|object
      */
-    public function findFieldValue(object $object, string $fieldName)
+    private function findFieldValue(object $object, string $fieldName)
     {
         $getter = 'get' . ucfirst($fieldName);
 
@@ -275,81 +207,5 @@ class Linker implements CompilerPassInterface
             || $item instanceof ClassDescriptor
             || $item instanceof TraitDescriptor
             || $item instanceof InterfaceDescriptor;
-    }
-
-    /**
-     * Replaces pseudo-types, such as `self`, into a normalized version based on the last container that was
-     * encountered.
-     *
-     * @todo can we remove the nullable from this somehow to make the method contents simpler
-     */
-    private function replacePseudoTypes(string $fqsen, ?DescriptorAbstract $container) : string
-    {
-        $pseudoTypes = ['self', '$this'];
-        foreach ($pseudoTypes as $pseudoType) {
-            if ((strpos($fqsen, $pseudoType . '::') !== 0 && $fqsen !== $pseudoType) || !$container) {
-                continue;
-            }
-
-            $fqsen = $container->getFullyQualifiedStructuralElementName()
-                . substr($fqsen, strlen($pseudoType));
-        }
-
-        return $fqsen;
-    }
-
-    /**
-     * Returns true if the context marker is found in the given FQSEN.
-     */
-    private function isContextMarkerInFqsen(string $fqsen) : bool
-    {
-        return strpos($fqsen, self::CONTEXT_MARKER) !== false;
-    }
-
-    /**
-     * Normalizes the given FQSEN as if the context marker represents a class/interface/trait as parent.
-     */
-    private function getTypeWithClassAsContext(string $fqsen, DescriptorAbstract $container) : string
-    {
-        if (!$container instanceof ClassDescriptor
-            && !$container instanceof InterfaceDescriptor
-            && !$container instanceof TraitDescriptor
-        ) {
-            return $fqsen;
-        }
-
-        $containerFqsen = $container->getFullyQualifiedStructuralElementName();
-
-        return str_replace(self::CONTEXT_MARKER . '::', $containerFqsen . '::', $fqsen);
-    }
-
-    /**
-     * Normalizes the given FQSEN as if the context marker represents a class/interface/trait as parent.
-     */
-    private function getTypeWithNamespaceAsContext(string $fqsen, DescriptorAbstract $container) : string
-    {
-        $namespace = $container instanceof NamespaceDescriptor ? $container : $container->getNamespace();
-        $fqnn = $namespace instanceof NamespaceDescriptor
-            ? $namespace->getFullyQualifiedStructuralElementName()
-            : $namespace;
-
-        return str_replace(self::CONTEXT_MARKER . '::', $fqnn . '\\', $fqsen);
-    }
-
-    /**
-     * Normalizes the given FQSEN as if the context marker represents the global namespace as parent.
-     */
-    private function getTypeWithGlobalNamespaceAsContext(string $fqsen) : string
-    {
-        return str_replace(self::CONTEXT_MARKER . '::', '\\', $fqsen);
-    }
-
-    /**
-     * Attempts to find an element with the given Fqsen in the list of elements for this project and returns null if
-     * it cannot find it.
-     */
-    private function fetchElementByFqsen(string $fqsen) : ?DescriptorAbstract
-    {
-        return $this->elementList[$fqsen] ?? null;
     }
 }
